Fixes + cleanup
- Added pytimemory.hpp for longer python functions
- fixed tuple issue with environment
- created timemory.hpp
- moved cxxtimemory.cpp to timemory.cpp
- moved non-namespace functions from manager to timemory.cpp
- f_field_list is now get_field_list() to help with static destruction
- fixed python 2.7 RSS test
- set TIMEMORY_TEST_MPI=OFF when MPI is not found
