+++ cmake/Modules/Packages.cmake
@@ -69,6 +69,7 @@ if(TIMEMORY_USE_MPI)
     else(MPI_FOUND)
 
         set(TIMEMORY_USE_MPI OFF)
+        set(TIMEMORY_TEST_MPI OFF)
         message(WARNING "MPI not found. Proceeding without MPI")
         remove_definitions(-DTIMEMORY_USE_MPI)
