+++ source/python/CMakeLists.txt
@@ -21,7 +21,8 @@ endif(NOT BUILD_SHARED_LIBS AND WIN32)
 #----------------------------------------------------------------------------
 # PyBind11
 #
-pybind11_add_module(${LIBNAME} ${CMAKE_CURRENT_LIST_DIR}/${LIBNAME}.cpp ${pybind_headers})
+pybind11_add_module(${LIBNAME} ${CMAKE_CURRENT_LIST_DIR}/${LIBNAME}.cpp
+    py${LIBNAME}.hpp ${pybind_headers})
 # link cxx library and external libraries
 target_link_libraries(${LIBNAME} PUBLIC ${LIBNAME}-cxx-library ${EXTERNAL_LIBRARIES})
 add_dependencies(${LIBNAME} ${LIBNAME}-c-library ${LIBNAME}-cxx-library)
