+++ source/timemory.cpp
@@ -29,6 +29,76 @@
 #include "timemory/manager.hpp"
 
 //============================================================================//
+// Easier to type than colons
+typedef std::thread::id thread_id_t;
+
+//============================================================================//
+// Helper function
+thread_id_t get_thread_id()
+{
+    return std::this_thread::get_id();
+}
+
+//============================================================================//
+//  initialized on main thread
+thread_id_t& _tim_manager_tid()
+{
+    tim_static_thread_local thread_id_t _instance_tid = get_thread_id();
+    return _instance_tid;
+}
+
+//============================================================================//
+//  initialized on main thread
+tim::manager*& _tim_manager_ptr()
+{
+    tim_static_thread_local tim::manager* _instance = nullptr;
+    return _instance;
+}
+
+//============================================================================//
+
+void _tim_manager_initialization()
+{
+    if(!_tim_manager_ptr())
+    {
+        _tim_manager_tid() = get_thread_id();
+        _tim_manager_ptr() = new tim::manager();
+    }
+    return;
+}
+
+//============================================================================//
+
+void _tim_manager_finalization()
+{
+    // this seems to fail on Travis but nowhere else...
+    if(get_thread_id() == _tim_manager_tid())
+    {
+        delete _tim_manager_ptr();
+        _tim_manager_ptr() = nullptr;
+    }
+    return;
+}
+
+//============================================================================//
+// These two functions are guaranteed to be called at load and
+// unload of the library containing this code.
+namespace
+{
+#if !defined(_WINDOWS)
+    void setup_timemory_manager(void) __attribute__ ((constructor));
+    void cleanup_timemory_manager(void) __attribute__((destructor));
+#endif
+    void setup_timemory_manager(void) { _tim_manager_initialization(); }
+    void cleanup_timemory_manager(void) { _tim_manager_finalization(); }
+}
+
+//============================================================================//
+
+CEREAL_CLASS_VERSION(tim::timer_tuple, TIMEMORY_TIMER_VERSION)
+CEREAL_CLASS_VERSION(tim::manager, TIMEMORY_TIMER_VERSION)
+
+//============================================================================//
 //
 //                      C++ interface
 //
